+++ source/data/index.test.js
@@ -105,6 +105,8 @@ const Data = t.struct({
     production: t.maybe(t.Boolean),
     staging: t.maybe(t.Boolean),
     build: t.maybe(t.Boolean),
+    buildSHA1: t.maybe(t.String),
+    buildNumber: t.maybe(t.String),
     hotModuleRloading: t.maybe(t.Boolean)
   })
 }, { name: 'Data' })