+++ source/scripts/main.js
@@ -1,9 +1,9 @@
 // Uncomment if you need to support advanced ES6 features in IE11 and below
 // import 'babel-polyfill'
 
-import { SITE } from '@data'
+import { SITE, ENV } from '@data'
 import './utils/dom-polyfills'
 
-console.log(`${SITE.name} v${SITE.version}`)
+console.log(`${SITE.name} v${SITE.version} #${ENV.buildNumber}, SHA1 ${ENV.buildSHA1}`)
 
 document.querySelector('html').classList.remove('no-js')